requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asciimatics==1.14.0
asgiref==3.6.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.4
django-bootstrap4==23.1
django-bootstrap5==22.2
django-grappelli==3.0.6
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
Faker==12.0.1
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
Flask==2.2.2
frozenlist==1.3.3
future==0.18.3
HTMLParser==0.0.2
idna==3.4
iniconfig==1.1.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.2.2
multidict==6.0.4
mypy==1.1.1
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.26.5
openpyxl==3.0.10
packaging==21.3
pep8-naming==0.13.3
Pillow==9.5.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyfiglet==0.8.post1
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pyTelegramBotAPI==4.7.1
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3
pywin32==306
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.12
wcwidth==0.2.6
Werkzeug==2.2.2
yarl==1.8.2
